Panic With Invalid Modification Time
If a file has a modification time before the unix epoch, insertion into
the files table will fail due to the CHECK constraint on fs_modified.

Check for an invalid modification time before trying to insert so we can
panic with a useful error containing the filename and modification time
instead of a generic "CHECK constraint failed: files" error.